Feat: 댓글 CRUD API 생성

+++ src/posts/posts.controller.ts
@@ -7,12 +7,14 @@ import {
   ParseIntPipe,
   Post,
   Put,
+  Query,
 } from '@nestjs/common';
-import { ApiBody, ApiTags } from '@nestjs/swagger';
+import { ApiBody, ApiQuery, ApiTags } from '@nestjs/swagger';
 import { PostsService } from './posts.service';
 import { CreatePostDTO } from './dto/create-post.dto';
 import { Posts } from './posts.entity';
 import { ModifyPostDTO } from './dto/modify-post.dto';
+import { FindManyPostDTO } from './dto/find-post.dto';
 
 // 토큰 적용 X 게시글 삭제, 수정 시 작성자인지 확인 해야 됨
 @ApiTags('posts')
@@ -27,8 +29,9 @@ export class PostsController {
   }
 
   @Get()
-  async getManyPosts() {
-    return await this.postService.getManyPost();
+  @ApiQuery({ type: FindManyPostDTO })
+  async getManyPosts(@Query() query: FindManyPostDTO) {
+    return await this.postService.getManyPost(query);
   }
 
   @Get(':id')
@@ -41,12 +44,17 @@ export class PostsController {
   async putPosts(
     @Param('id', ParseIntPipe) id: number,
     @Body() body: ModifyPostDTO,
+    user_id: number,
   ) {
-    return await this.postService.modifyPost({ where: { id }, data: body });
+    return await this.postService.modifyPost({
+      where: { id },
+      data: body,
+      user_id,
+    });
   }
 
   @Delete(':id')
-  async deletePost(@Param('id', ParseIntPipe) id: number) {
-    return await this.postService.removePost({ id });
+  async deletePost(@Param('id', ParseIntPipe) id: number, user_id: number) {
+    return await this.postService.removePost({ where: { id }, user_id });
   }
 }
